4_app/ehs-splunk-demo/requirements.txt
azure-eventhub==5.0.0b4
fluent-logger
splunk_handler
azure.keyvault
